Added path tiles and created a path layout.
Tiles hold whether it is a path, whether it has a tower on top or not, and the position.
The path manager creates the layout of the path and puts cubes onto the screen. These cubes are green if they are the path or brown if they are just normal tiles.